refactor/simplifications
